Fix last warnings

+++ lib/include/detfct.h
@@ -23,11 +23,11 @@ namespace GF2_Utils {
 	   Determinant est egale aux produits des elements de la diagonale de la triangularisation.
 	   Des qu'un element nul est rencontre sur la diagonale alors retourne zero.
 	   */
-	bool det_b(std::vector<std::bitset<MAT_MAX_SIZE>> M, long msize);
+	bool det_b(std::vector<std::bitset<MAT_MAX_SIZE>> M, size_t msize);
 
-	bool chk_triangular_tables_not_equal(std::vector<std::bitset<MAT_MAX_SIZE>> & M1, std::vector<std::bitset<MAT_MAX_SIZE>> & M2, long maxdim, long msize);
+	bool chk_triangular_tables_not_equal(std::vector<std::bitset<MAT_MAX_SIZE>> & M1, std::vector<std::bitset<MAT_MAX_SIZE>> & M2, long maxdim, size_t msize);
 
-	int how_many_differences(std::vector<std::bitset<MAT_MAX_SIZE>> & M1, std::vector<std::bitset<MAT_MAX_SIZE>> & M2, long maxdim, long msize);
+	int how_many_differences(std::vector<std::bitset<MAT_MAX_SIZE>> & M1, std::vector<std::bitset<MAT_MAX_SIZE>> & M2, long maxdim, size_t msize);
 }
 
 #endif
